Moves Gameboard into component
